Manambahkan fitur sidebar, client dan contact person

+++ app/Http/Controllers/SizeController.php
@@ -15,18 +15,15 @@ class SizeController extends Controller
      */
     public function index(): Response
     {
-        // $users = User::with('sizes')->get();
-        // $sizes = Size::with('user')->get();
         return response()-> view ('dashboard.media.sizes.index', [
-            'sizes'=>Size::with(['user'])->get(),
+            'sizes'=>Size::filter(request('search'))->sortable()->with(['user'])->orderBy("size", "asc")->paginate(10)->withQueryString(),
             'title' => 'Daftar Ukuran'
-            // compact('sizes', 'users')
         ]);
     }
 
     public function showSize(){
-        $dataSize = Size::All();
-
+        $dataSize = Size::orderBy("size", "asc")->get();
+        
         return response()->json(['dataSize'=> $dataSize]);
     }
 
@@ -35,7 +32,13 @@ class SizeController extends Controller
      */
     public function create(): Response
     {
-        //
+        if(auth()->user()->level === 'Administrator' || auth()->user()->level === 'Media'){
+            return response()-> view ('dashboard.media.sizes.create', [
+                'title' => 'Menambahkan Ukuran'
+            ]);
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -43,7 +46,28 @@ class SizeController extends Controller
      */
     public function store(Request $request): RedirectResponse
     {
-        //
+        if(auth()->user()->level === 'Administrator' || auth()->user()->level === 'Media'){
+            if($request->side == '0'){
+                return back()->withErrors(['side' => ['Silahkan pilih jumlah sisi']])->withInput();
+            }
+
+            if($request->orientation == 'Pilih Orientasi'){
+                return back()->withErrors(['orientation' => ['Silahkan pilih orientasi']])->withInput();
+            }
+        
+            $validateData = $request->validate([
+                'size' => 'required',
+                'side' => 'required',
+                'orientation' => 'required'
+            ]);
+    
+            $validateData['user_id'] = auth()->user()->id;
+            Size::create($validateData);
+    
+            return redirect('/dashboard/media/sizes')->with('success','Ukuran '. $request->size . ' berhasil ditambahkan');
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -51,7 +75,10 @@ class SizeController extends Controller
      */
     public function show(Size $size): Response
     {
-        //
+        return response()-> view ('dashboard.media.sizes.show', [
+            'size' => $size,
+            'title' => 'Detail Ukuran ' . $size->size
+        ]);
     }
 
     /**
@@ -59,7 +86,14 @@ class SizeController extends Controller
      */
     public function edit(Size $size): Response
     {
-        //
+        if(auth()->user()->level === 'Administrator' || auth()->user()->level === 'Media'){
+            return response()->view('dashboard.media.sizes.edit', [
+                'size' => $size,
+                'title' => 'Edit Ukuran'
+            ]);
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -67,7 +101,34 @@ class SizeController extends Controller
      */
     public function update(Request $request, Size $size): RedirectResponse
     {
-        //
+        if(auth()->user()->level === 'Administrator' || auth()->user()->level === 'Media'){
+            if ($request->size == $size->size && $request->side == $size->side && $request->orientation == $size->orientation ) {
+                return redirect('/dashboard/media/sizes')->with('success','Ukuran '. $request->size . ' tidak ada perubahan');
+            } else {
+                if($request->side == '0'){
+                    return back()->withErrors(['side' => ['Silahkan pilih jumlah sisi']])->withInput();
+                }
+
+                if($request->orientation == 'Pilih Orientasi'){
+                    return back()->withErrors(['orientation' => ['Silahkan pilih orientasi']])->withInput();
+                }
+            
+                $validateData = $request->validate([
+                    'size' => 'required',
+                    'side' => 'required',
+                    'orientation' => 'required'
+                ]);
+        
+                $validateData['user_id'] = auth()->user()->id;
+                
+                Size::where('id', $size->id)
+                    ->update($validateData);
+        
+                return redirect('/dashboard/media/sizes')->with('success','Ukuran '. $size->size . ' berhasil diupdate');
+            }
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -75,8 +136,12 @@ class SizeController extends Controller
      */
     public function destroy(Size $size): RedirectResponse
     {
-        Size::destroy($size->id);
+        if(auth()->user()->level === 'Administrator' || auth()->user()->level === 'Media'){
+            Size::destroy($size->id);
 
-        return redirect('/dashboard/media/sizes')->with('success','Ukuran '. $size->size . $size->side .' sisi '. $size->orientation .' berhasil dihapus');
+            return redirect('/dashboard/media/sizes')->with('success','Ukuran '. $size->size .' berhasil dihapus');
+        } else {
+            abort(403);
+        }
     }
 }
